Three tier product e2e tests start on the three tier landing page
Also includes some refactoring to move common code to complete the
generic checkout form to utils.
